Starting to move the logic into separete class and out of the forms
