+++ UrlaubsPlaner/Main_Form.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+using UrlaubsPlaner.DBInteraction;
+using UrlaubsPlaner.Entities;
+
+namespace UrlaubsPlaner
+{
+    public partial class Main_Form : Form
+    {
+        public Main_Form()
+        {
+            InitializeComponent();
+        }
+    }
+}
