+++ UrlaubsPlaner/Program.cs
@@ -17,7 +17,7 @@ namespace UrlaubsPlaner
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form_Main());
+            Application.Run(new Main_Form());
         }
     }
 }
