+++ UrlaubsPlaner/UrlaubsPlaner.csproj
@@ -63,6 +63,7 @@
     <Compile Include="AbsenceType_Form.Designer.cs">
       <DependentUpon>AbsenceType_Form.cs</DependentUpon>
     </Compile>
+    <Compile Include="Controller\Main_FormController.cs" />
     <Compile Include="DBInteraction\AbsenceQuerys.cs" />
     <Compile Include="DBInteraction\AbsenceTypeQuerys.cs" />
     <Compile Include="DBInteraction\CountryQuerys.cs" />
@@ -82,11 +83,11 @@
     <Compile Include="Entities\EntityTransformations.cs" />
     <Compile Include="Entities\IEntity.cs" />
     <Compile Include="Entities\IUpsertable.cs" />
-    <Compile Include="Form_Main.cs">
+    <Compile Include="Main_Form.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form_Main.Designer.cs">
-      <DependentUpon>Form_Main.cs</DependentUpon>
+    <Compile Include="Main_Form.Designer.cs">
+      <DependentUpon>Main_Form.cs</DependentUpon>
     </Compile>
     <Compile Include="DBInteraction\DataBaseConnection.cs" />
     <Compile Include="Program.cs" />
@@ -97,8 +98,8 @@
     <EmbeddedResource Include="Employee_Form.resx">
       <DependentUpon>Employee_Form.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Form_Main.resx">
-      <DependentUpon>Form_Main.cs</DependentUpon>
+    <EmbeddedResource Include="Main_Form.resx">
+      <DependentUpon>Main_Form.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
